requirements.txt
beautifulsoup4
requests
pdf2image
pytesseract
python-dotenv
openai
jsonschema